accounts(checks): refactor to top level module
