+++ rdmo/accounts/tests/test_checks_shibboleth.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from ..checks.shibboleth import W_FIX_DISABLED, W_MISSING_FIX, check_shibboleth_remoteuser
+from rdmo.accounts.checks import W_FIX_DISABLED, W_MISSING_FIX, check_shibboleth_remoteuser
+
 from .helpers import enable_shibboleth, fake_shibboleth  # noqa: F401
